organize file structure of webview
