+++ packages/zowe-explorer/src/webviews/src/edit-history/components/PersistentClearAllButton.tsx
@@ -1,0 +1,20 @@
+import { VSCodeButton } from "@vscode/webview-ui-toolkit/react";
+import { JSXInternal } from "preact/src/jsx";
+import PersistentVSCodeAPI from "./PersistentVSCodeAPI";
+
+export default function PersistentClearAllButton({ type }: { type: string }): JSXInternal.Element {
+  const handleClick = () => {
+    PersistentVSCodeAPI.getVSCodeAPI().postMessage({
+      command: "clear-all",
+      attrs: {
+        type,
+      },
+    });
+  };
+
+  return (
+    <VSCodeButton style={{ maxWidth: "20vw" }} onClick={handleClick}>
+      Clear All
+    </VSCodeButton>
+  );
+}
