+++ packages/zowe-explorer/src/webviews/src/edit-history/components/PersistentRefreshButton.tsx
@@ -1,0 +1,20 @@
+import { VSCodeButton } from "@vscode/webview-ui-toolkit/react";
+import { JSXInternal } from "preact/src/jsx";
+import PersistentVSCodeAPI from "./PersistentVSCodeAPI";
+
+export default function PersistentRefreshButton({ type }: { type: string }): JSXInternal.Element {
+  const handleClick = () => {
+    PersistentVSCodeAPI.getVSCodeAPI().postMessage({
+      command: "refresh",
+      attrs: {
+        type,
+      },
+    });
+  };
+
+  return (
+    <VSCodeButton style={{ maxWidth: "20vw", marginRight: "15px" }} onClick={handleClick}>
+      Refresh
+    </VSCodeButton>
+  );
+}
