+++ packages/zowe-explorer/src/webviews/src/edit-history/components/PersistentVSCodeAPI.ts
@@ -1,0 +1,9 @@
+import { WebviewApi } from "vscode-webview";
+
+export default class PersistentVSCodeAPI {
+    private static vscodeAPI = acquireVsCodeApi();
+
+    public static getVSCodeAPI(): WebviewApi<unknown> {
+        return this.vscodeAPI;
+    }
+}
